Add custom Task and Subtask validators
Created custom validators for these two entities because there are cases where simply adding @Valid and constraints on entity fields is not enough.

Example: when updating a task, a user can update the title, or the description, or both. With NotBlank and Size constraints on both fields, an error occurs if a user tries to update only one field.

The custom validators allow one or both fields to be updated.

Refactored Task and Subtask DTOs to add custom validator annotation and remove previous annotations.