Refactor the update board method
Previous version only updated the board name, which did not meet update needs. The refactored version updates the board and associated columns, including adding or removing columns from a board.

This utilizes a newly created column service method to perform column updates.

Also removed the 'getBoardById' method as it was not used.