Refactor the column service
Removed the 'add' and 'delete' column methods as they're now controlled in the 'update' method.

The 'update' method is called from the board service layer and adds, updates, and/or deletes columns and returns them to the boards method.

Also includes a new custom query to the column repository to support the 'update' method.